requirements.txt
discord.py
motor
mercadopago
openpyxl